requirements.txt
Adafruit-Blinka==8.45.2
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-requests==4.1.3
adafruit-circuitpython-typing==1.10.3
adafruit-circuitpython-vl53l0x==3.6.11
Adafruit-PlatformDetect==3.71.0
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.8.1
annotated-types==0.7.0
args==0.1.0
beautifulsoup4==4.12.3
binho-host-adapter==0.1.6
blinker==1.8.2
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
clint==0.5.1
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
distro==1.5.0
flask==3.0.3
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
gpiozero==1.6.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==2.10
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
Markdown==3.6
markdownify==0.13.1
MarkupSafe==2.1.5
numpy==2.0.0
pathfinding==1.0.10
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.3
pydantic-core==2.18.4
pyftdi==0.55.4
pygame==2.6.0
pyinotify==0.9.6
pyOpenSSL==20.0.1
pyparsing==3.1.2
pyserial==3.5
python-apt==2.2.1
python-prctl==1.7
pyusb==1.2.1
pyxattr==0.7.2
requests==2.25.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
simplejpeg==1.7.4
simplejson==3.17.2
six==1.16.0
soupsieve==2.5
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
toml==0.10.1
tqdm==4.66.4
typing-extensions==4.12.1
uritemplate==4.1.1
urllib3==1.26.5
v4l2-python3==0.3.2
werkzeug==3.0.3
youtube-dl==2021.6.6
zipp==3.19.1
